# api/api.yaml
openapi: 3.1.0
info:
  title: Polygon ID
  description: |
    Documentation for the Backend Verifier API
  version: "1"

servers:
  - description: Local
    url: http://localhost:3009

tags:
  - name: Verifier
    description: Collection of endpoints related to Verifier

paths:
  /:
    get:
      summary: Get the documentation
      operationId: GetDocumentation
      x-internal: true
      responses:
        200:
          description: success and returns the documentation in HTML format

  /health:
    get:
      summary: Health Check
      operationId: Health
      responses:
        '200':
          description: All services are running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
        '500':
          $ref: '#/components/responses/500'

  /sign-in:
    get:
      summary: Sing in
      operationId: Sign in
      tags:
        - Verifier
      requestBody:
        content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignInRequest'
      responses:
        '200':
          description: All services are running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingInResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'

  /status:
    get:
      summary: Get Status
      operationId: Status
      tags:
        - Verifier
      parameters:
          - $ref: '#/components/parameters/sessionID'
      responses:
        '200':
          description: Get response status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'

  /qr-store:
    post:
      summary: Store QRCode
      operationId: QRStore
      tags:
        - Verifier
      requestBody:
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/QRStoreRequest'
      responses:
        '200':
          description: QR Indirection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QRStoreResponse'
        '500':
          $ref: '#/components/responses/500'
    get:
      summary: Get QRCode from store
      operationId: GetQRCodeFromStore
      tags:
        - Verifier
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: QR Code Indirection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QRCode'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'

  /callback:
    post:
      summary: Callback
      operationId: Callback
      tags:
        - Verifier
      parameters:
        - $ref: '#/components/parameters/sessionID'
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              example: jwz-token
      responses:
        '200':
          description: Callback
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallbackResponse'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'

components:
  schemas:
    CallbackResponse:
      type: object

    QRStoreResponse:
      type: string
      example: iden3comm://?request_uri=https%3A%2F%2Fissuer-demo.polygonid.me%2Fapi%2Fqr-store%3Fid%3Df780a169-8959-4380-9461-f7200e2ed3f4

    QRStoreRequest:
      $ref: '#/components/schemas/QRCode'
      example: ' {
       "id":"6dc645a6-2be3-4099-a645-20784ee53cd0",
       "typ":"application/iden3comm-plain-json",
       "type":"https://iden3-communication.io/authorization/1.0/request",
       "thid":"6dc645a6-2be3-4099-a645-20784ee53cd0",
       "body":{
          "callbackUrl":"https://self-hosted-demo-backend-platform.polygonid.me/api/callback?sessionId=469345",
          "reason":"test flow",
          "scope":[
             {
                "id":1,
                "circuitId":"credentialAtomicQuerySigV2",
                "query":{
                   "allowedIssuers":[
                      "*"
                   ],
                   "context":"https://raw.githubusercontent.com/iden3/claim-schema-vocab/main/schemas/json-ld/kyc-v4.jsonld",
                   "credentialSubject":{
                      "birthday":{
                         "$lt":20000101
                      }
                   },
                   "skipClaimRevocationCheck":true,
                   "type":"KYCAgeCredential"
                }
             }
          ]
       },
       "from":"did:polygonid:polygon:mumbai:2qH7TstpRRJHXNN4o49Fu9H2Qismku8hQeUxDVrjqT"
      } '


    StatusResponse:
      type: object
      example: '
      {
   "authResponse":{
      "id":"0e3f5aa0-6d7a-4199-a40d-861541140c1f",
      "typ":"application/iden3-zkp-json",
      "type":"https://iden3-communication.io/authorization/1.0/response",
      "thid":"b328c339-3ed2-4312-a85a-8bb238fa9809",
      "body":{
         "did_doc":{
            "context":[
               "https://www.w3.org/ns/did/v1"
            ],
            "id":"did:polygonid:polygon:mumbai:2qF1bpKZ8R2MVVq9GwQRB54J1qSZneSGGzm8GhFk6G",
            "service":[
               {
                  "id":"did:polygonid:polygon:mumbai:2qF1bpKZ8R2MVVq9GwQRB54J1qSZneSGGzm8GhFk6G#push",
                  "type":"push-notification",
                  "serviceEndpoint":"https://push-staging.polygonid.com/api/v1",
                  "metadata":{
                     "devices":[
                        {
                           "ciphertext":"KwZwhskHTscYkD5NPNHV8WgQN1Rtwvzww3YnAgGtPcalzyKDXYUI9XH8ChbNdcw7Lxa4W269mHO5ZDlIfQgCaM78sH5edaDabvCDyNDE/7jBn/RsNz1GJ2/Ke2nFCp1j950eQuO41zEr5LOIDj9pCLMAXcco28brnIrFFIxJ8u/dxBkmgbzNCqFJnxebsULVcONylNUGWB79z2xaMuoW6BiigfB8R2F9AvfBRt13+VjJQaLpBz77KHSmwwqZBglGgCdLIq16ysqfP2z1S83yVma3vgbMWfHj36LPiDueV29L1KVREFEtouVOhaTeKd8gDHDluERWjbbgPCpCa4fLNvP1i+bVY6PkmK1A1o2yxcZQ+8ynJ+Scm+erovTB82qYCNyJwHUddluSdgCdijLZXwMnBF30jULGhVhlsJSTe1bzovnejVM0mxTRSGJ/kxaQsIWUi022QZeGx2I5zlxmof+SYfwQk9Vs/DFV1GRO8xbzoAeeKU8lbeethGgtDVSXltZ97oJpH4zknSM2LYmrYPZ1L00gLLXoSK6JYLge6aiFIR2gf2Im4CoPkAc0laR06DBXoaTXF53t5VPl6G5q9dVoKvWzzF6chnkaIgFuhT0A+c0smhzeq+7PU190q0KxZj9bkPQF04CpAISg1OAUntEmCCFikxQaw4xtv2fo4qY=",
                           "alg":"RSA-OAEP-512"
                        }
                     ]
                  }
               }
            ]
         },
         "scope":[
            {
               "id":1,
               "circuitId":"credentialAtomicQuerySigV2",
               "proof":{
                  "pi_a":[
                     "12916887140854457892647620502091853527223284508643597174268220900460474475161",
                     "7815885961037450817114852798791437723324816574329366451705470564398354226204",
                     "1"
                  ],
                  "pi_b":[
                     [
                        "14625934917855489647326046307562857961424756237253021317135922833820611194046",
                        "5018745463002238259245574426431170022254742695230343890143233419076586082008"
                     ],
                     [
                        "9777610290700415716437972808701853631866043782554140794025362158572330126401",
                        "3432191018521184021144084259408894556250671674562773286025197254477859374498"
                     ],
                     [
                        "1",
                        "0"
                     ]
                  ],
                  "pi_c":[
                     "6039342622666338645512510363735849552178457866691174435937389399241189306246",
                     "19750246574434214606174786191432199098816211892852326627851987022804413689469",
                     "1"
                  ],
                  "protocol":"groth16"
               },
               "pub_signals":[
                  "1",
                  "24517555279028353051178157460399931994339755856689989623157395286174208514",
                  "19510132996725366596454568784662601562609350327133164810707533894557776854012",
                  "1",
                  "21196425925540611834072050424245907522377653852612334807514247180354327042",
                  "1",
                  "10793476797837836072392758215441493304137919984792281022340864278228859852306",
                  "1702633337",
                  "106228138578177371556412734740259405073",
                  "0",
                  "8262158451645466116825601889503011908601155175273968686900962251943496227039",
                  "0",
                  "1",
                  "1",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0",
                  "0"
               ]
            }
         ]
      },
      "from":"did:polygonid:polygon:mumbai:2qF1bpKZ8R2MVVq9GwQRB54J1qSZneSGGzm8GhFk6G",
      "to":"did:polygonid:polygon:mumbai:2qH7TstpRRJHXNN4o49Fu9H2Qismku8hQeUxDVrjqT"
   },
   "id":"did:polygonid:polygon:mumbai:2qF1bpKZ8R2MVVq9GwQRB54J1qSZneSGGzm8GhFk6G",
   "jwz":"eyJhbGciOiJncm90aDE2IiwiY2lyY3VpdElkIjoiYXV0aFYyIiwiY3JpdCI6WyJjaXJjdWl0SWQiXSwidHlwIjoiYXBwbGljYXRpb24vaWRlbjMtemtwLWpzb24ifQ.eyJpZCI6IjBlM2Y1YWEwLTZkN2EtNDE5OS1hNDBkLTg2MTU0MTE0MGMxZiIsInR5cCI6ImFwcGxpY2F0aW9uL2lkZW4zLXprcC1qc29uIiwidHlwZSI6Imh0dHBzOi8vaWRlbjMtY29tbXVuaWNhdGlvbi5pby9hdXRob3JpemF0aW9uLzEuMC9yZXNwb25zZSIsInRoaWQiOiJiMzI4YzMzOS0zZWQyLTQzMTItYTg1YS04YmIyMzhmYTk4MDkiLCJib2R5Ijp7ImRpZF9kb2MiOnsiY29udGV4dCI6WyJodHRwczovL3d3dy53My5vcmcvbnMvZGlkL3YxIl0sImlkIjoiZGlkOnBvbHlnb25pZDpwb2x5Z29uOm11bWJhaToycUYxYnBLWjhSMk1WVnE5R3dRUkI1NEoxcVNabmVTR0d6bThHaEZrNkciLCJzZXJ2aWNlIjpbeyJpZCI6ImRpZDpwb2x5Z29uaWQ6cG9seWdvbjptdW1iYWk6MnFGMWJwS1o4UjJNVlZxOUd3UVJCNTRKMXFTWm5lU0dHem04R2hGazZHI3B1c2giLCJ0eXBlIjoicHVzaC1ub3RpZmljYXRpb24iLCJzZXJ2aWNlRW5kcG9pbnQiOiJodHRwczovL3B1c2gtc3RhZ2luZy5wb2x5Z29uaWQuY29tL2FwaS92MSIsIm1ldGFkYXRhIjp7ImRldmljZXMiOlt7ImNpcGhlcnRleHQiOiJLd1p3aHNrSFRzY1lrRDVOUE5IVjhXZ1FOMVJ0d3Z6d3czWW5BZ0d0UGNhbHp5S0RYWVVJOVhIOENoYk5kY3c3THhhNFcyNjltSE81WkRsSWZRZ0NhTTc4c0g1ZWRhRGFidkNEeU5ERS83akJuL1JzTnoxR0oyL0tlMm5GQ3Axajk1MGVRdU80MXpFcjVMT0lEajlwQ0xNQVhjY28yOGJybklyRkZJeEo4dS9keEJrbWdiek5DcUZKbnhlYnNVTFZjT055bE5VR1dCNzl6MnhhTXVvVzZCaWlnZkI4UjJGOUF2ZkJSdDEzK1ZqSlFhTHBCejc3S0hTbXd3cVpCZ2xHZ0NkTElxMTZ5c3FmUDJ6MVM4M3lWbWEzdmdiTVdmSGozNkxQaUR1ZVYyOUwxS1ZSRUZFdG91Vk9oYVRlS2Q4Z0RIRGx1RVJXamJiZ1BDcENhNGZMTnZQMWkrYlZZNlBrbUsxQTFvMnl4Y1pRKzh5bkorU2NtK2Vyb3ZUQjgycVlDTnlKd0hVZGRsdVNkZ0NkaWpMWlh3TW5CRjMwalVMR2hWaGxzSlNUZTFiem92bmVqVk0wbXhUUlNHSi9reGFRc0lXVWkwMjJRWmVHeDJJNXpseG1vZitTWWZ3UWs5VnMvREZWMUdSTzh4YnpvQWVlS1U4bGJlZXRoR2d0RFZTWGx0Wjk3b0pwSDR6a25TTTJMWW1yWVBaMUwwMGdMTFhvU0s2SllMZ2U2YWlGSVIyZ2YySW00Q29Qa0FjMGxhUjA2REJYb2FUWEY1M3Q1VlBsNkc1cTlkVm9Ldld6ekY2Y2hua2FJZ0Z1aFQwQStjMHNtaHplcSs3UFUxOTBxMEt4Wmo5YmtQUUYwNENwQUlTZzFPQVVudEVtQ0NGaWt4UWF3NHh0djJmbzRxWT0iLCJhbGciOiJSU0EtT0FFUC01MTIifV19fV19LCJtZXNzYWdlIjpudWxsLCJzY29wZSI6W3sicHJvb2YiOnsicGlfYSI6WyIxMjkxNjg4NzE0MDg1NDQ1Nzg5MjY0NzYyMDUwMjA5MTg1MzUyNzIyMzI4NDUwODY0MzU5NzE3NDI2ODIyMDkwMDQ2MDQ3NDQ3NTE2MSIsIjc4MTU4ODU5NjEwMzc0NTA4MTcxMTQ4NTI3OTg3OTE0Mzc3MjMzMjQ4MTY1NzQzMjkzNjY0NTE3MDU0NzA1NjQzOTgzNTQyMjYyMDQiLCIxIl0sInBpX2IiOltbIjE0NjI1OTM0OTE3ODU1NDg5NjQ3MzI2MDQ2MzA3NTYyODU3OTYxNDI0NzU2MjM3MjUzMDIxMzE3MTM1OTIyODMzODIwNjExMTk0MDQ2IiwiNTAxODc0NTQ2MzAwMjIzODI1OTI0NTU3NDQyNjQzMTE3MDAyMjI1NDc0MjY5NTIzMDM0Mzg5MDE0MzIzMzQxOTA3NjU4NjA4MjAwOCJdLFsiOTc3NzYxMDI5MDcwMDQxNTcxNjQzNzk3MjgwODcwMTg1MzYzMTg2NjA0Mzc4MjU1NDE0MDc5NDAyNTM2MjE1ODU3MjMzMDEyNjQwMSIsIjM0MzIxOTEwMTg1MjExODQwMjExNDQwODQyNTk0MDg4OTQ1NTYyNTA2NzE2NzQ1NjI3NzMyODYwMjUxOTcyNTQ0Nzc4NTkzNzQ0OTgiXSxbIjEiLCIwIl1dLCJwaV9jIjpbIjYwMzkzNDI2MjI2NjYzMzg2NDU1MTI1MTAzNjM3MzU4NDk1NTIxNzg0NTc4NjY2OTExNzQ0MzU5MzczODkzOTkyNDExODkzMDYyNDYiLCIxOTc1MDI0NjU3NDQzNDIxNDYwNjE3NDc4NjE5MTQzMjE5OTA5ODgxNjIxMTg5Mjg1MjMyNjYyNzg1MTk4NzAyMjgwNDQxMzY4OTQ2OSIsIjEiXSwicHJvdG9jb2wiOiJncm90aDE2IiwiY3VydmUiOiJibjEyOCJ9LCJwdWJfc2lnbmFscyI6WyIxIiwiMjQ1MTc1NTUyNzkwMjgzNTMwNTExNzgxNTc0NjAzOTk5MzE5OTQzMzk3NTU4NTY2ODk5ODk2MjMxNTczOTUyODYxNzQyMDg1MTQiLCIxOTUxMDEzMjk5NjcyNTM2NjU5NjQ1NDU2ODc4NDY2MjYwMTU2MjYwOTM1MDMyNzEzMzE2NDgxMDcwNzUzMzg5NDU1Nzc3Njg1NDAxMiIsIjEiLCIyMTE5NjQyNTkyNTU0MDYxMTgzNDA3MjA1MDQyNDI0NTkwNzUyMjM3NzY1Mzg1MjYxMjMzNDgwNzUxNDI0NzE4MDM1NDMyNzA0MiIsIjEiLCIxMDc5MzQ3Njc5NzgzNzgzNjA3MjM5Mjc1ODIxNTQ0MTQ5MzMwNDEzNzkxOTk4NDc5MjI4MTAyMjM0MDg2NDI3ODIyODg1OTg1MjMwNiIsIjE3MDI2MzMzMzciLCIxMDYyMjgxMzg1NzgxNzczNzE1NTY0MTI3MzQ3NDAyNTk0MDUwNzMiLCIwIiwiODI2MjE1ODQ1MTY0NTQ2NjExNjgyNTYwMTg4OTUwMzAxMTkwODYwMTE1NTE3NTI3Mzk2ODY4NjkwMDk2MjI1MTk0MzQ5NjIyNzAzOSIsIjAiLCIxIiwiMSIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCIsIjAiLCIwIiwiMCJdLCJpZCI6MSwiY2lyY3VpdElkIjoiY3JlZGVudGlhbEF0b21pY1F1ZXJ5U2lnVjIifV19LCJmcm9tIjoiZGlkOnBvbHlnb25pZDpwb2x5Z29uOm11bWJhaToycUYxYnBLWjhSMk1WVnE5R3dRUkI1NEoxcVNabmVTR0d6bThHaEZrNkciLCJ0byI6ImRpZDpwb2x5Z29uaWQ6cG9seWdvbjptdW1iYWk6MnFIN1RzdHBSUkpIWE5ONG80OUZ1OUgyUWlzbWt1OGhRZVV4RFZyanFUIn0.eyJwcm9vZiI6eyJwaV9hIjpbIjE2MDY2Mzc4ODgyMjA4MTkzMjg3MDkzNzQxMjE3MDUyMjU0NzkxODgwMTg2MzE0MjMxMTU5MDI2MTczMzI5OTkzODczMDk1MTA0NzgwIiwiMTkyOTI4MjgwMzI5MzcyNzczOTk5MDU2OTY4MDAzMzA3NDY3MzAzMTYyOTMyNDU0NzY2NjA2NTk0Mjc1NTU5NzczMjY4OTU1MzI1MTgiLCIxIl0sInBpX2IiOltbIjE2MTE2ODY0NTc2MDg5NDQ5NzY4NDI5MDg5NjE5ODEyODk4NDQ0ODQwMDMwMTE1MjU5NjEwNzE5MTc1Nzc0MTIxNDEyMTM2NTI0OTQ2IiwiNzY4MzYzMzc3MjY2MjY3OTM0NjM3Nzc0NzYxNzU5NDg0MjgzOTM4OTI2MDUzMzcyNDQ4NDQzMTY5MDkzOTM1OTQxMjc5ODI4MTU0Il0sWyIxODY4NzQ3ODU2Mzk4OTQ2NjMzMDUwNjQyMjc3Nzc1MTM4NTY5NTY4MDk4NjMyNjY4NjEwMTY5NjQ5MDY4MDg3NTgzNTIyMTk0NjU0NiIsIjEwMzY1MjMwNDIxOTAxNTI3NDgwMzM0MTUwMTMyMDk5NzI0MTc2NDMxNDg2NTcyNzExMDI4NTQ3MDAyMzQ4NzQ0MTUwNDI4Nzc2OTY4Il0sWyIxIiwiMCJdXSwicGlfYyI6WyIyMTE4NDU4NDU3NTM2NTQ2MDIzMjY0ODc4NTk5Nzg1MzQ1Mjc4Njg5MzEzNDY5MTU3MzI3Nzc4MDI2NzU3NzQ0MDcxMTgyODgzNzYyNSIsIjYzNjY5NjgxOTQ1OTAzNTk3Mjc5ODczMTYxNjU5MTUyMjEzMTU5MTAxNzI2NDM1ODcwMzc4MDc3NzY2MTUyNjk1ODgxMjkyMTUwNjMiLCIxIl0sInByb3RvY29sIjoiZ3JvdGgxNiIsImN1cnZlIjoiYm4xMjgifSwicHViX3NpZ25hbHMiOlsiMjQ1MTc1NTUyNzkwMjgzNTMwNTExNzgxNTc0NjAzOTk5MzE5OTQzMzk3NTU4NTY2ODk5ODk2MjMxNTczOTUyODYxNzQyMDg1MTQiLCIxNjA3MjY1NzAyMjIxODcxMTM3NjYzNTEzNDg3NjkxNzUyODAzOTk3OTA1MjA0NTI4MjIzNDE0MjA4ODMyOTgyNjIxNzUwNDE0MDQzNSIsIjQyMTc4MjI2NjU0MzM4MDcyMjg1MjY0NjU4MTE4MTU3Nzk1OTk3ODczMTc3Mzk1NTYxODc2Nzg1ODkxNjM3ODI4ODkzMjQ2MTU5ODQiXX0",
   "proofs":[
      {
         "id":1,
         "circuitId":"credentialAtomicQuerySigV2",
         "proof":{
            "pi_a":[
               "12916887140854457892647620502091853527223284508643597174268220900460474475161",
               "7815885961037450817114852798791437723324816574329366451705470564398354226204",
               "1"
            ],
            "pi_b":[
               [
                  "14625934917855489647326046307562857961424756237253021317135922833820611194046",
                  "5018745463002238259245574426431170022254742695230343890143233419076586082008"
               ],
               [
                  "9777610290700415716437972808701853631866043782554140794025362158572330126401",
                  "3432191018521184021144084259408894556250671674562773286025197254477859374498"
               ],
               [
                  "1",
                  "0"
               ]
            ],
            "pi_c":[
               "6039342622666338645512510363735849552178457866691174435937389399241189306246",
               "19750246574434214606174786191432199098816211892852326627851987022804413689469",
               "1"
            ],
            "protocol":"groth16"
         },
         "pub_signals":[
            "1",
            "24517555279028353051178157460399931994339755856689989623157395286174208514",
            "19510132996725366596454568784662601562609350327133164810707533894557776854012",
            "1",
            "21196425925540611834072050424245907522377653852612334807514247180354327042",
            "1",
            "10793476797837836072392758215441493304137919984792281022340864278228859852306",
            "1702633337",
            "106228138578177371556412734740259405073",
            "0",
            "8262158451645466116825601889503011908601155175273968686900962251943496227039",
            "0",
            "1",
            "1",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0",
            "0"
         ]
      }
   ]
}'

    Health:
      type: object
      x-omitempty: false
      example:
        healthy: true

    SingInResponse:
      type: object
      required:
          - sessionID
          - qrCode
      properties:
          sessionID:
            type: integer
          qrCode:
            $ref: '#/components/schemas/QRCode'

    QRCode:
      type: object
      required:
          - id
          - typ
          - type
          - thid
          - body
          - from
      properties:
        id:
          type: string
          example: '6dc645a6-2be3-4099-a645-20784ee53cd0'
        typ:
          type: string
          example: 'application/iden3comm-plain-json'
        type:
          type: string
          example: 'https://iden3-communication.io/authorization/1.0/request'
        thid:
          type: string
          example: '6dc645a6-2be3-4099-a645-20784ee53cd0'
        body:
          type: object
          properties:
            callbackUrl:
              type: string
              example: 'https://self-hosted-demo-backend-platform.polygonid.me/api/callback?sessionId=469345'
            reason:
              type: string
              example: 'test flow'
            scope:
              type: array
              items:
                $ref: '#/components/schemas/Scope'
        from:
            type: string
            example: 'did:polygonid:polygon:mumbai:2qH7TstpRRJHXNN4o49Fu9H2Qismku8hQeUxDVrjqT'
        to:
          type: string
          example: 'did:polygon:mumbai:2qJ689kpoJxcSzB5sAFJtPsSBSrHF5dq722BHMqURL'

    Scope:
      type: object
      required:
        - id
        - circuitId
        - query
      properties:
        id:
          type: integer
          example: 1
        circuitId:
          type: string
          example: 'credentialAtomicQuerySigV2'
        query:
          type: object

    SignInRequest:
      type: object
      required:
        - network
        - circuitID
        - requestID
        - query
      properties:
        network:
          type: string
          example: 'mumbai'
        circuitID:
          type: string
          example: 'credentialAtomicQuerySigV2'
        requestID:
          type: integer
          example: 1
        to:
          type: string
          example: 'did:xxxx'
        query:
          type: object
          example: '
          {
             "type":"KYCAgeCredential",
             "context":"https://raw.githubusercontent.com/iden3/claim-schema-vocab/main/schemas/json-ld/kyc-v3.jsonld",
             "credentialSubject":{
                "birthday":{
                   "$lt":20000101
                }
             },
             "skipClaimRevocationCheck":true,
             "allowedIssuers":[
                "*"
             ]
          }'
        onChain:
          $ref: '#/components/schemas/OnChain'

    OnChain:
      type: object
      required:
          - network
          - contractAddress
          - chainID
          - methodIDs
      properties:
          network:
            type: string
            example: 'mumbai'
          contractAddress:
            type: string
            example: '0x000000'
          methodID:
            type: string
            example: 'whatever'
          chainID:
            type: integer
            example: 80001


    UUID:
      type: string
      x-go-type: uuid.UUID
      x-go-type-import:
        name: uuid
        path: github.com/google/uuid
      example: 8edd8112-c415-11ed-b036-debe37e1cbd6

    GenericErrorMessage:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: 'Something happen'


  parameters:
    sessionID:
      name: sessionID
      in: query
      required: true
      description: |
        Session ID e.g: 123456
      schema:
        type: string


    id:
      name: id
      in: query
      required: true
      description: |
        ID e.g: 89d298fa-15a6-4a1d-ab13-d1069467eedd
      schema:
        type: string
        x-go-type: uuid.UUID
        x-go-type-import:
          name: uuid
          path: github.com/google/uuid

  responses:
    '400':
      description: 'Bad Request'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorMessage'

    '404':
      description: 'Not Found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorMessage'

    '500':
      description: 'Internal Server error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorMessage'